infound-backend-services/pyproject.toml
[project]
name = "infound-backend-services"
version = "0.1.0"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi (>=0.124.0,<0.125.0)",
    "uvicorn[standard] (>=0.38.0)",
    "pydantic (>=2.12.5)",
    "python-multipart (>=0.0.7)",
    "httpx (>=0.27.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic-settings (>=2.12.0)",
    "pyjwt (>=2.8.0)",
    "redis (>=7.1.0,<8.0.0)",
    "passlib[bcrypt] (>=1.7.4)",
    "structlog (>=24.3.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "concurrent-log-handler (>=0.9.28,<0.10.0)",
    "uuid6 (>=2025.0.1,<2026.0.0)",

    "ruamel-yaml (>=0.18.16,<0.19.0)",
    "aiomysql (>=0.3.2,<0.4.0)",
    "greenlet (>=3.0.3,<4.0.0)",
    "aio-pika (>=9.3.0,<10.0.0)"
]

# Python dependency groups (PEP 631, dev deps)
[dependency-groups]
dev = [
    "flake8 (>=7.3.0)",
    "black (>=25.11.0)",
    "poetry-core (>=2.2.1,<3.0.0)",
    "sqlacodegen (>=3.2.0,<4.0.0)",
]

[tool.poetry]
# Disable package mode
package-mode = false

# Build system config
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
